+++ plugins/org.eclipse.osee.framework.skynet.core/src/org/eclipse/osee/framework/skynet/core/httpRequests/CommitBranchHttpRequestOperation.java
@@ -148,7 +148,7 @@ public final class CommitBranchHttpRequestOperation extends AbstractOperation {
                   change.getModificationType().isDeleted() ? RelationEventType.Deleted : change.getModificationType().isUnDeleted() ? RelationEventType.Undeleted : RelationEventType.Added;
 
                DefaultBasicIdRelation defaultBasicGuidRelation = new DefaultBasicIdRelation(relChange.getBranch(),
-                  relChange.getRelationType().getId(), relChange.getItemId().getId().intValue(), relChange.getGamma(),
+                  relChange.getRelationType().getId(), relChange.getItemId().getId(), relChange.getGamma(),
                   relChange.getChangeArtifact().getBasicGuidArtifact(),
                   relChange.getEndTxBArtifact().getBasicGuidArtifact());
                EventBasicGuidRelation event = new EventBasicGuidRelation(relationEventType, relChange.getArtId(),
